Move e2e to e2e-new
