+++ test/e2e/capv_quick_start_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package e2enew
+package e2e
 
 import (
 	"context"
