--- test/e2e/capv_test_utils.go
@@ -1,30 +1,0 @@
-/*
-Copyright 2020 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package e2e
-
-import (
-	"crypto/sha256"
-	"fmt"
-
-	"github.com/google/uuid"
-)
-
-// Hash7 returns a random, seven character string.
-func Hash7() string {
-	randomUUID := uuid.New()
-	return fmt.Sprintf("%x", sha256.Sum256(randomUUID[:]))[:7]
-}
