+++ test/e2e/data/cni/calico/calico.yaml
