+++ test/e2e/data/cni/kindnet/kindnet.yaml
