--- test/e2e/e2e.conf
@@ -1,73 +1,0 @@
----
-images:
-- name: capv-manager:e2e
-  loadBehavior: mustLoad
-- name: gcr.io/k8s-staging-cluster-api/cluster-api-controller:master
-  loadBehavior: tryLoad
-- name: gcr.io/k8s-staging-cluster-api/kubeadm-bootstrap-controller:master
-  loadBehavior: tryLoad
-- name: gcr.io/k8s-staging-cluster-api/kubeadm-control-plane-controller:master
-  loadBehavior: tryLoad
-
-kubernetesVersion: v1.18.2
-
-components:
-
-# Load the certificate manager and wait for all of its pods and service to
-# become available.
-- name:    cert-manager
-  sources:
-  - type:  url
-    value: https://github.com/jetstack/cert-manager/releases/download/v0.11.1/cert-manager.yaml
-  waiters:
-  - type:  service
-    value: v1beta1.webhook.cert-manager.io
-  - value: cert-manager
-
-# Load CAPI core and wait for its pods to become available.
-- name:    capi
-  sources:
-  - value: https://github.com/kubernetes-sigs/cluster-api//config?ref=master
-    replacements:
-    - old: "imagePullPolicy: Always"
-      new: "imagePullPolicy: IfNotPresent"
-    - old: "--feature-gates=MachinePool=\\${EXP_MACHINE_POOL:=false},ClusterResourceSet=\\${EXP_CLUSTER_RESOURCE_SET:=false}"
-      new: "--feature-gates=MachinePool=false,ClusterResourceSet=false"
-  waiters:
-  - value: capi-system
-
-# Load the CAPI kubeadm bootstrapper and wait for its pods to become available.
-- name:    capi-kubeadm-bootstrap
-  sources:
-  - value: https://github.com/kubernetes-sigs/cluster-api//bootstrap/kubeadm/config?ref=master
-    replacements:
-    - old: "imagePullPolicy: Always"
-      new: "imagePullPolicy: IfNotPresent"
-    - old: "--feature-gates=MachinePool=\\${EXP_MACHINE_POOL:=false}"
-      new: "--feature-gates=MachinePool=false"
-  waiters:
-  - value: capi-kubeadm-bootstrap-system
-
-# Load the CAPI kubeadm control plane and wait for its pods to become available.
-- name:    capi-kubeadm-control-plane
-  sources:
-  - value: https://github.com/kubernetes-sigs/cluster-api//controlplane/kubeadm/config?ref=master
-    replacements:
-    - old: "imagePullPolicy: Always"
-      new: "imagePullPolicy: IfNotPresent"
-    - old: "--feature-gates=MachinePool=\\${EXP_MACHINE_POOL:=false}"
-      new: "--feature-gates=MachinePool=false"
-  waiters:
-  - value: capi-kubeadm-control-plane-system
-
-# Load CAPV and wait for its pods to become available.
-- name:    capv
-  sources:
-  - value: ../../config
-    replacements:
-    - old: gcr.io/cluster-api-provider-vsphere/release/manager:latest
-      new: capv-manager:e2e
-    - old: "imagePullPolicy: Always"
-      new: "imagePullPolicy: IfNotPresent"
-  waiters:
-  - value: capv-system
