+++ test/e2e/govmomi_test.go
@@ -17,30 +17,34 @@ limitations under the License.
 package e2e
 
 import (
+	"context"
 	"flag"
-	"fmt"
 	"net/url"
 	"os"
-	"strings"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
 	"github.com/vmware/govmomi"
 	"github.com/vmware/govmomi/find"
-	"github.com/vmware/govmomi/object"
 	"github.com/vmware/govmomi/vim25/soap"
+)
+
+var (
+	vsphereUsername = os.Getenv("VSPHERE_USERNAME")
+	vspherePassword = os.Getenv("VSPHERE_PASSWORD")
+
+	vsphereServer     string
+	vsphereDatacenter string
+
+	ctx           = context.TODO()
+	vsphereClient *govmomi.Client
+	vsphereFinder *find.Finder
 )
 
 func init() {
 	flag.StringVar(&vsphereServer, "e2e.vsphereServer", os.Getenv("VSPHERE_SERVER"), "the vSphere server used for e2e tests")
 	flag.StringVar(&vsphereDatacenter, "e2e.vsphereDataceter", os.Getenv("VSPHERE_DATACENTER"), "the inventory path of the vSphere datacenter in which VMs are created")
-	flag.StringVar(&vsphereFolder, "e2e.vsphereFolder", os.Getenv("VSPHERE_FOLDER"), "the inventory path of the vSphere folder in which VMs are created")
-	flag.StringVar(&vspherePool, "e2e.vspherePool", os.Getenv("VSPHERE_RESOURCE_POOL"), "the inventory path of the vSphere resource pool in which VMs are created")
-	flag.StringVar(&vsphereDatastore, "e2e.vsphereDatastore", os.Getenv("VSPHERE_DATASTORE"), "the name of the vSphere datastore in which VMs are created")
-	flag.StringVar(&vsphereNetwork, "e2e.vsphereNetwork", os.Getenv("VSPHERE_NETWORK"), "the name of the vSphere network to which VMs are connected")
-	flag.StringVar(&vsphereMachineTemplate, "e2e.vsphereMachineTemplate", os.Getenv("VSPHERE_MACHINE_TEMPLATE"), "the template from which the Kubernetes VMs are cloned")
-	flag.StringVar(&vsphereHAProxyTemplate, "e2e.vsphereHAProxyTemplate", os.Getenv("VSPHERE_HAPROXY_TEMPLATE"), "the template from which the HAProxy load balancer VM is cloned")
 }
 
 func initVSphereSession() {
@@ -62,26 +66,4 @@ func initVSphereSession() {
 	datacenter, err := vsphereFinder.DatacenterOrDefault(ctx, vsphereDatacenter)
 	Expect(err).ShouldNot(HaveOccurred())
 	vsphereFinder.SetDatacenter(datacenter)
-}
-
-func destroyVMsWithPrefix(prefix string) {
-	vmList, _ := vsphereFinder.VirtualMachineList(ctx, vspherePool)
-	for _, vm := range vmList {
-		if strings.HasPrefix(vm.Name(), prefix) {
-			destroyVM(vm)
-		}
-	}
-}
-
-func destroyVM(vm *object.VirtualMachine) {
-	if task, _ := vm.PowerOff(ctx); task != nil {
-		if err := task.Wait(ctx); err != nil {
-			fmt.Printf("error powering off %s machine: %s\n", vm.Name(), err)
-		}
-	}
-	if task, _ := vm.Destroy(ctx); task != nil {
-		if err := task.Wait(ctx); err != nil {
-			fmt.Printf("error destroying  %s machine: %s\n", vm.Name(), err)
-		}
-	}
 }
