--- test/e2enew/common.go
@@ -1,32 +1,0 @@
-/*
-Copyright 2020 The Kubernetes Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package e2enew
-
-import (
-	"fmt"
-
-	. "github.com/onsi/ginkgo" // // nolint:golint,stylecheck
-)
-
-// Test suite constants for e2e config variables
-const (
-	KubernetesVersion = "KUBERNETES_VERSION"
-)
-
-func Byf(format string, a ...interface{}) {
-	By(fmt.Sprintf(format, a...))
-}
